Restore original blog test and add view-the-blog acceptance test in interested-partner-organization
