+++ core/tests/puppeteer-acceptance-tests/specs/logged-out-user/view-and-search-blog-posts.spec.ts
@@ -13,145 +13,77 @@
 // limitations under the License.
 
 /**
- * @fileoverview Acceptance test from CUJv3 Doc
- * https://docs.google.com/document/d/1D7kkFTzg3rxUe3QJ_iPlnxUzBFNElmRkmAWss00nFno/
- *
- * IO.2. View the blog.
+ * @fileoverview Acceptance Test for viewing and searching blog posts as a logged-out user.
  */
 
 import {UserFactory} from '../../utilities/common/user-factory';
 import {LoggedOutUser} from '../../utilities/user/logged-out-user';
-import {BlogPostEditor} from '../../utilities/user/blog-post-editor';
-import {LoggedInUser} from '../../utilities/user/logged-in-user';
+import {SuperAdmin} from '../../utilities/user/super-admin';
 import testConstants from '../../utilities/common/test-constants';
+import {showMessage} from '../../utilities/common/show-message';
 
 const DEFAULT_SPEC_TIMEOUT_MSECS = testConstants.DEFAULT_SPEC_TIMEOUT_MSECS;
-const ROLES = testConstants.Roles;
+const NUM_OF_DUMMY_BLOGS = 30;
 
 describe('Logged-out User', function () {
-  let interestParent: LoggedOutUser;
-  let blogPostWriter: BlogPostEditor & LoggedInUser;
+  let loggedOutUser: LoggedOutUser;
+  let superAdmin: SuperAdmin;
 
   beforeAll(async function () {
-    interestParent = await UserFactory.createLoggedOutUser();
-    blogPostWriter = await UserFactory.createNewUser(
-      'blogPostWriter',
-      'blog_post_writer@example.com',
-      [ROLES.BLOG_POST_EDITOR]
-    );
+    loggedOutUser = await UserFactory.createLoggedOutUser();
+    superAdmin = await UserFactory.createNewSuperAdmin('superAdm');
+
+    const isInProdMode = await superAdmin.isInProdMode();
+    if (!isInProdMode) {
+      await superAdmin.navigateToAdminPageActivitiesTab();
+      await superAdmin.generateDummyBlogPosts(NUM_OF_DUMMY_BLOGS);
+    } else {
+      showMessage(
+        'The application is currently running in production mode. Skipping dummy blog post generation.'
+      );
+      // TODO(#22716): Complete it with Milestone 2.5 Implement Acceptance
+      // tests for Blog Admin and Blog Post Editor.
+      process.exit(0);
+    }
+  }, DEFAULT_SPEC_TIMEOUT_MSECS);
+
+  beforeEach(async function () {
+    await loggedOutUser.navigateToBlogPage();
   }, DEFAULT_SPEC_TIMEOUT_MSECS);
 
   it(
-    'should display the welcome message and empty-state notice when no blog posts exist',
+    'should be able to view blog posts with tags',
     async function () {
-      await interestParent.navigateToBlogPageViaNavbar();
-      await interestParent.expectBlogWelcomeMessageToBeVisible();
-      await interestParent.expectNoBlogPostsMessageToBeVisible();
+      await loggedOutUser.expectBlogPostsToHaveAtLeastOneTag();
     },
     DEFAULT_SPEC_TIMEOUT_MSECS
   );
 
   it(
-    'should not display draft blog posts to logged-out users',
+    'should be able to navigate through blog pages using pagination',
     async function () {
-      // As blogPostWriter: Create and publish blog "Blog 1".
-      await blogPostWriter.navigateToBlogDashboardPage();
-      await blogPostWriter.openBlogEditorPage();
-      await blogPostWriter.uploadBlogPostThumbnailImage();
-      await blogPostWriter.updateBlogPostTitle('Blog 1');
-      await blogPostWriter.updateBodyTextTo('Content for Blog 1');
-      await blogPostWriter.selectTag('News');
-      await blogPostWriter.saveBlogBodyChanges();
-      await blogPostWriter.publishTheBlogPost();
+      await loggedOutUser.expectBlogPaginationControlsVisible();
 
-      await blogPostWriter.navigateToBlogDashboardPage();
-      await blogPostWriter.openBlogEditorPage();
-      await blogPostWriter.uploadBlogPostThumbnailImage();
-      await blogPostWriter.updateBlogPostTitle('Blog 2');
-      await blogPostWriter.updateBodyTextTo('Content for Blog 2');
-      await blogPostWriter.selectTag('News');
-      await blogPostWriter.saveBlogBodyChanges();
-      await blogPostWriter.saveTheDraftBlogPost();
-
-      // As interestParent: Refresh the page.
-      await interestParent.navigateToBlogPage();
-
-      await interestParent.expectNumberOfBlogPostsOnPageToBe(1);
-      await interestParent.expectBlogPostWithTitleToBePresent('Blog 1');
+      await loggedOutUser.clickNextBlogPage();
+      await loggedOutUser.clickPreviousBlogPage();
     },
     DEFAULT_SPEC_TIMEOUT_MSECS
   );
 
   it(
-    'should paginate blog posts after the first 10 items',
+    'should be able to search blog posts by tags',
     async function () {
-      // As blogPostWriter: Create and publish 11 new blogs.
-      await blogPostWriter.navigateToBlogDashboardPage();
-      for (let i = 3; i <= 13; i++) {
-        await blogPostWriter.openBlogEditorPage();
-        await blogPostWriter.uploadBlogPostThumbnailImage();
-        await blogPostWriter.updateBlogPostTitle(`Blog ${i}`);
-        await blogPostWriter.updateBodyTextTo(`Content for Blog ${i}`);
-        await blogPostWriter.selectTag('News');
-        await blogPostWriter.saveBlogBodyChanges();
-        await blogPostWriter.publishTheBlogPost();
-        await blogPostWriter.navigateToBlogDashboardPage();
-      }
-      // Pagination shows 10 blogs per page. Since 12 blogs exist,
-      // page 1 should display 10 blogs and page 2 should display the remaining 2.
-      await interestParent.navigateToBlogPage();
-      await interestParent.expectNumberOfBlogPostsOnPageToBe(10);
-
-      await interestParent.clickNextBlogPage();
-      await interestParent.expectNumberOfBlogPostsOnPageToBe(2);
+      await loggedOutUser.filterBlogPostsByTag('Community');
+      await loggedOutUser.expectBlogSearchResultsToHaveTag('Community');
     },
     DEFAULT_SPEC_TIMEOUT_MSECS
   );
 
   it(
-    'should display blog page with correct layout elements',
+    'should be able to search blog posts by keywords',
     async function () {
-      await interestParent.navigateToBlogPageViaNavbar();
-
-      await interestParent.expectBlogPageLayoutToBeCorrect();
-    },
-    DEFAULT_SPEC_TIMEOUT_MSECS
-  );
-
-  it(
-    'should display blog post page with all required elements',
-    async function () {
-      await interestParent.navigateToBlogPage();
-      await interestParent.clickOnFirstBlogPost();
-
-      await interestParent.expectBlogPostPageElementsToBeVisible();
-    },
-    DEFAULT_SPEC_TIMEOUT_MSECS
-  );
-
-  it(
-    'should be able to search and filter blog posts',
-    async function () {
-      await blogPostWriter.navigateToBlogDashboardPage();
-      await blogPostWriter.openBlogEditorPage();
-      await blogPostWriter.uploadBlogPostThumbnailImage();
-      await blogPostWriter.updateBlogPostTitle('International Blog');
-      await blogPostWriter.updateBodyTextTo('International News Blog');
-      await blogPostWriter.selectTag('International');
-      await blogPostWriter.saveBlogBodyChanges();
-      await blogPostWriter.publishTheBlogPost();
-
-      await interestParent.navigateToBlogPage();
-      await interestParent.filterBlogPostsByTag('International');
-      await interestParent.expectBlogSearchResultsToHaveTag('International');
-
-      await interestParent.navigateToBlogPage();
-      await interestParent.filterBlogPostsByKeyword('International');
-      await interestParent.expectBlogSearchResultsToContain('International');
-
-      await interestParent.navigateToBlogPage();
-      await interestParent.filterBlogPostsByKeyword('educators');
-      await interestParent.expectNoBlogPostsMessageToBeVisible();
+      await loggedOutUser.filterBlogPostsByKeyword('Education');
+      await loggedOutUser.expectBlogSearchResultsToContain('Education');
     },
     DEFAULT_SPEC_TIMEOUT_MSECS
   );
